refactor to use external CSS files
