+++ package.json
@@ -33,6 +33,7 @@
   "devDependencies": {
     "@greenwood/cli": "~0.31.0",
     "@greenwood/plugin-adapter-vercel": "~0.31.0",
+    "@greenwood/plugin-import-raw": "~0.31.0",
     "@greenwood/plugin-renderer-lit": "~0.31.0",
     "@greenwood/plugin-typescript": "~0.31.0",
     "rimraf": "^6.0.1"
