+++ src/layouts/app.html
@@ -10,8 +10,8 @@
 
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Source+Sans+Pro">
     <link rel="stylesheet" href="../styles/main.css">
-    <script type="module" src="../components/card.ts"></script>
-    <script type="module" src="../components/modal.ts"></script>
+    <script type="module" src="../components/card/card.ts"></script>
+    <script type="module" src="../components/modal/modal.ts"></script>
   </head>
 
   <body>
