+++ src/types.d.ts
@@ -1,0 +1,5 @@
+// Greenwood Raw plugin
+declare module "*?type=raw" {
+  const content: string;
+  export default content;
+}
